Unwrap Info-directory-list from unless
In terminal emacs, Info-directory-list automatically points to
Info-default-directory-list, so the expected behavior occurs. But in
Aquamacs (and I suspect other windowed emacsen), calling info only
shows the Org manual and no others. I believe this is because
Info-directory list is nil but not void. The (unless ... ) call always
returns nil, so Info-directory-list ends up containing only
"~/.emacs.d/src/org/doc".

By unwrapping the variable definition, we *unconditionally* copy
Info-default-directory-list to Info-directory-list, which may not be
desirable. I do not know what it was wrapped in the first place, so
this may not be the best solution, but the change produces the
expected effect on my system.

GNU Emacs 23.2.50.1 (i386-apple-darwin9.8.0, NS apple-appkit-949.54)
of 2010-08-18 on braeburn.aquamacs.org - Aquamacs Distribution 2.1

Org-mode version 7.4 (release_7.4.127.gec97)